Add files via upload
Hei sorry for the mess and not doing libraries and all of that, this is just functions added on top of the data parser but i think it's quite clear what it is doing and i only had a moment to implement it. Anyway this thing gets rid of the less important axis and fits an ellipse on the projection. They do not really look like ellipses (is that because of irregular sampling or just because they are not ellipses?). 

If you have python 2 you might have to change .__next__ to .next
You also need to move the sync in patient folder 1 out of the most inner vcg folder